requirements.txt
Flask==3.0.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
nltk==3.8.1
spacy==3.7.2
requests==2.31.0
python-dotenv==1.0.0
elasticsearch==8.11.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pyyaml==6.0.1
